[Refactor] LottoAutoController 생성 후 비지니스 로직 분리, LottoAuto 메서드 분리
